requirements.txt
pandas==2.1.4
snowflake-connector-python==3.6.0
python-dotenv==1.0.0
pyarrow==14.0.2
matplotlib
seaborn
